New tools and cron file modifications (#487)
