--- tools/Cron_jobs/msre_reload.crond
@@ -1,21 +1,0 @@
-# $Id: msre_reload.crond,v 1.2 2004/10/13 01:29:51 jofcore Exp $
-# 
-# msre = MailScanner Ruleset Editor
-# (c) 2004 Kevin Hanser
-# Released under the GNU GPL: http://www.gnu.org/copyleft/gpl.html#TOC1
-# 
-#    This program is free software; you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation; either version 2 of the License, or
-#    (at your option) any later version.
-# 
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
-#    along with this program; if not, write to the Free Software
-#    Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
-#
-*/5 * * * *	root	/usr/local/bin/msre_reload.sh
