--- tools/Cron_jobs/quarantine_maint.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-/usr/local/bin/quarantine_maint.php --clean
