+++ tools/MailScanner_rule_editor/msre_reload.crontab
@@ -1,0 +1,6 @@
+# MailWatch for MailScanner
+# Copyright (C) 2003-2011  Steve Freegard (steve@freegard.name)
+# Copyright (C) 2011  Garrod Alwood (garrod.alwood@lorodoes.com)
+# Copyright (C) 2014-2017  MailWatch Team (https://github.com/mailwatch/1.2.0/graphs/contributors)
+#
+*/5 * * * *	root	/usr/local/bin/msre_reload.sh
