+++ tools/Postfix_relay/INSTALL
@@ -1,14 +1,46 @@
 ----------------------------------
 INSTALLATION INSTRUCTIONS
 ----------------------------------
-----------------------------------------------------------
-Optional for item Postfix
-----------------------------------------------------------
 
------ Adding Postfix relay information
+Optional files for Postfix MTA
 
-- Edit the parser (mailwatch_relay.sh) and add it as an hourly cron job
+Copy the 'mailwatch-postfix-relay' cron file into /etc/cron.hourly/
 
-- Edit the parser for location of MailWatch webpages
+Copy in /usr/local/bin the following files:
 
-- Edit mailscanner_relay.php in the mailscanner folder
+mailwatch_postfix_relay.php
+mailwatch_mailscanner_relay.php
+
+Verify if the 2 files are executable. If not, use "chmod +x" to correct tis problem.
+
+Note about mailwatch_postfix_relay.php: 
+
+The script have two options, run it once per hour with '--refresh' or tail the log continuously.
+
+The 'mailwatch-postfix-relay' cron use the '--refresh' method.
+
+Example of script for the 'tail' method.
+
+----------------------------------
+#!/bin/bash
+
+# @asuweb
+
+if ps -C php -o args h | grep mailwatch_postfix_relay.php
+then true ## mailwatch_postfix_relay.php running
+else /usr/bin/php -q /usr/local/bin/mailwatch_postfix_relay.php > /dev/null &
+fi
+if ps -C php -o args h | grep mailwatch_mailscanner_relay.php
+then true ## mailwatch_mailscanner_relay.php running
+else /usr/bin/php -q /usr/local/bin/mailwatch_mailscanner_relay.php > /dev/null &
+fi
+----------------------------------
+
+And you run it from cron every hour.
+
+ATTENTION:
+
+Between 1.0.3 or 1.2.0 RCx version many files used by MailWatch have changed of name 
+and changed of place.
+Particularly the PHP files that are launched in cron and the Init script files.
+Please read the documentation carefully to remove old files during the upgrade.
