+++ tools/Postfix_relay/mailwatch-postfix-relay
@@ -1,0 +1,14 @@
+#!/bin/bash
+
+# MailWatch for MailScanner
+# Copyright (C) 2003-2011  Steve Freegard (steve@freegard.name)
+# Copyright (C) 2011  Garrod Alwood (garrod.alwood@lorodoes.com)
+# Copyright (C) 2014-2017  MailWatch Team (https://github.com/mailwatch/1.2.0/graphs/contributors)
+
+test -x /usr/local/bin/mailwatch_postfix_relay.php || exit 0
+/usr/bin/php -q /usr/local/bin/mailwatch_postfix_relay.php --refresh
+
+test -x /usr/local/bin/mailwatch_mailscanner_relay.php || exit 0
+/usr/bin/php -q /usr/local/bin/mailwatch_mailscanner_relay.php --refresh
+
+exit 0
