+++ tools/Sendmail-Exim_queue/INSTALL
@@ -1,0 +1,25 @@
+-----------------------------
+INSTALLATION INSTRUCTIONS
+-----------------------------
+
+sendmail_queue.php process the MTA Exim or Sendmail queue to store messages in MailWatch database and see them in the MailWatch GUI.
+
+Copy the mailwatch_sendmail_queue.php file in /usr/local/bin
+
+Verify if the file is executable. If not, use "chmod +x" to correct tis problem.
+
+Edit CRONTAB:
+
+# crontab -e
+
+# Run sendmail_queue.php each minute
+0-59 * * * * 	/usr/local/bin/mailwatch_sendmail_queue.php
+
+Note: mailwatch_sendmail_queue.php re-creates all entries on each run, so for busy sites you will probably want to change this to run every 5 minutes or greater.
+
+ATTENTION:
+
+Between 1.0.3 or 1.2.0 RCx version many files used by MailWatch have changed of name 
+and changed of place.
+Particularly the PHP files that are launched in cron and the Init script files.
+Please read the documentation carefully to remove old files during the upgrade.
