+++ tools/Sendmail-Exim_queue/mailq.crontab
@@ -1,0 +1,7 @@
+# MailWatch for MailScanner
+# Copyright (C) 2003-2011  Steve Freegard (steve@freegard.name)
+# Copyright (C) 2011  Garrod Alwood (garrod.alwood@lorodoes.com)
+# Copyright (C) 2014-2017  MailWatch Team (https://github.com/mailwatch/1.2.0/graphs/contributors)
+#
+# Run mailq.php each minute
+0-59 * * * * 	/usr/local/bin/mailq.php
