--- tools/Sendmail_queue/INSTALL
@@ -1,13 +1,0 @@
------------------------------
-INSTALLATION INSTRUCTIONS
------------------------------
-
------------ Setting up MailWatch to see the sendmail queue
-
-Edit mailq.php first to change the require line to point to the location of functions.php, then:
-
-# cp tools/Sendmail_queue/mailq.php /usr/local/bin
-# crontab -e
-0-59 * * * * 	/usr/local/bin/mailq.php
-
-Note: mailq.php re-creates all entries on each run, so for busy sites you will probably want to change this to run every 5 minutes or greater.
