+++ tools/Sendmail_relay/INSTALL
@@ -1,19 +1,27 @@
+----------------------------------
 INSTALLATION INSTRUCTIONS
--------------------------
+----------------------------------
 
-- Setup the Sendmail Relay Log watcher
+Setup for the Sendmail Relay Log watcher:
 
-You can get MailWatch to watch your sendmail logs and store all message relay information which is then displayed on the 'Message Detail' page which helps debugging and makes it easy for a Helpdesk to actually see where a message was delivered to by the MTA and what the response back was (e.g. the remote queue id etc.).
+You can get MailWatch to watch your Sendmail MTA logs and store all message relay information which is then displayed on the 'Message Detail' page which helps debugging and makes it easy for a Helpdesk to actually see where a message was delivered to by the MTA and what the response back was (e.g. the remote queue id etc.).
 
 On Debian/Ubuntu:
 
-# cp tools/Sendmail_relay/sendmail_relay.php /usr/local/bin/.
-# cp tools/Sendmail_relay/sendmail_relay.init /etc/init.d/.
-# chmod +x /usr/local/bin/sendmail_relay.php
-# chmod +x /etc/init.d/sendmail_relay
-# /etc/init.d/sendmail_relay start
-# update-rc.d sendmail_relay defaults
+# cp tools/Sendmail_relay/mailwatch_sendmail_relay.php /usr/local/bin/.
+# cp tools/Sendmail_relay/mailwatch-sendmail-relay /etc/init.d/.
+# chmod +x /usr/local/bin/mailwatch_sendmail_relay.php
+# chmod +x /etc/init.d/mailwatch-sendmail-relay
+# /etc/init.d/mailwatch-sendmail-relay start
+# update-rc.d mailwatch-sendmail-relay defaults
 
-For others distributiikns, please change according to.
+For others Linux distributions, please change according to.
 
-ATTENTION: By default, sendmail_relay.php run with root user. Change user to your Apache or MTA user and check right of /var/log/mail.log.
+By default, mailwatch_sendmail_relay.php run with 'root' user. Change user to your Apache or Sendmail MTA user (check right of /var/log/mail.log).
+
+ATTENTION:
+
+Between 1.0.3 or 1.2.0 RCx version many files used by MailWatch have changed of name 
+and changed of place.
+Particularly the PHP files that are launched in cron and the Init script files.
+Please read the documentation carefully to remove old files during the upgrade.
